move project finder methods to separate module
